refactor: :art: adds constants folder to hold message and command constants
