+++ packages/stacked_tools/lib/src/mixins/project_structure_validator_mixin.dart
@@ -1,6 +1,6 @@
 import 'package:stacked_tools/src/exceptions/invalid_stacked_structure_exception.dart';
 import 'package:stacked_tools/src/locator.dart';
-import 'package:stacked_tools/src/message_constants.dart';
+import 'package:stacked_tools/src/constants/message_constants.dart';
 import 'package:stacked_tools/src/services/file_service.dart';
 
 mixin ProjectStructureValidator {
